# Additional Perf. Optimizations