--- a/src/VCDiff.Benchmark/Program.cs
+++ b/src/VCDiff.Benchmark/Program.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using BenchmarkDotNet.Running;
+
+namespace VCDiff.Benchmark
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            BenchmarkRunner.Run<RandomDataDecode>();
+            BenchmarkRunner.Run<RandomDataEncode>();
+        }
+    }
+}