--- a/src/VCDiff.Benchmark/RandomDataEncode.cs
+++ b/src/VCDiff.Benchmark/RandomDataEncode.cs
@@ -5,6 +5,7 @@ using VCDiff.Encoders;
 
 namespace VCDiff.Benchmark
 {
+    [MemoryDiagnoser()]
     [SimpleJob(1, 3, 8)]
     public class RandomDataEncode
     {