--- a/src/VCDiff.Benchmark/VCDiff.Benchmark.csproj
+++ b/src/VCDiff.Benchmark/VCDiff.Benchmark.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\VCDiff\VCDiff.csproj" />
+  </ItemGroup>
+
+</Project>