--- a/src/VCDiff.sln
+++ b/src/VCDiff.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.29721.120
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VCDiff", "VCDiff\VCDiff.csproj", "{4D676F4E-E66D-4A41-861F-600F0FFCD052}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VCDiff.Tests", "VCDiff.Tests\VCDiff.Tests.csproj", "{724A8CFC-F566-47BB-ABE7-2A7DF5EC4E66}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VCDiff.Tests", "VCDiff.Tests\VCDiff.Tests.csproj", "{724A8CFC-F566-47BB-ABE7-2A7DF5EC4E66}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VCDiff.Benchmark", "VCDiff.Benchmark\VCDiff.Benchmark.csproj", "{29FF95D4-B9BF-480B-B0F1-E5A0BAEAAA82}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{724A8CFC-F566-47BB-ABE7-2A7DF5EC4E66}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{724A8CFC-F566-47BB-ABE7-2A7DF5EC4E66}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{724A8CFC-F566-47BB-ABE7-2A7DF5EC4E66}.Release|Any CPU.Build.0 = Release|Any CPU
+		{29FF95D4-B9BF-480B-B0F1-E5A0BAEAAA82}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{29FF95D4-B9BF-480B-B0F1-E5A0BAEAAA82}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{29FF95D4-B9BF-480B-B0F1-E5A0BAEAAA82}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{29FF95D4-B9BF-480B-B0F1-E5A0BAEAAA82}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE