--- a/src/VCDiff/Decoders/BodyDecoder.cs
+++ b/src/VCDiff/Decoders/BodyDecoder.cs
@@ -179,13 +179,13 @@ namespace VCDiff.Decoders
         /// Decode if as expecting interleave
         /// </summary>
         /// <returns></returns>
-        public VCDiffResult DecodeInterleave()
+        public Task<VCDiffResult> DecodeInterleave()
         {
             var result = DecodeInterleaveCore();
             targetData.Seek(0, SeekOrigin.Begin);
             targetData.CopyTo(outputStream);
             targetData.SetLength(0);
-            return result;
+            return Task.FromResult(result);
         }
 
         /// <summary>
@@ -272,13 +272,13 @@ namespace VCDiff.Decoders
         /// Decode normally
         /// </summary>
         /// <returns></returns>
-        public VCDiffResult Decode()
+        public Task<VCDiffResult> Decode()
         {
             var result = this.DecodeCore();
             targetData.Seek(0, SeekOrigin.Begin);
             targetData.CopyTo(outputStream);
             targetData.SetLength(0);
-            return result;
+            return Task.FromResult(result);
         }
 
         /// <summary>