… src/VCDiff/Decoders/WindowDecoder.cs
… src/VCDiff/Encoders/BlockHash.cs
… src/VCDiff/Encoders/ChunkEncoder.cs
… src/VCDiff/Encoders/InstructionMap.cs
… src/VCDiff/Encoders/RollingHash.cs
… src/VCDiff/Encoders/VcEncoder.cs
… src/VCDiff/Encoders/WindowEncoder.cs
… src/VCDiff/Shared/AddressCache.cs
… src/VCDiff/Shared/Adler32.cs
… src/VCDiff/Shared/ByteBuffer.cs
… src/VCDiff/Shared/ByteStreamReader.cs
… src/VCDiff/Shared/Checksum.cs
… src/VCDiff/Shared/CodeTable.cs
… src/VCDiff/Shared/Extensions.cs
… src/VCDiff/Shared/IByteBuffer.cs
… src/VCDiff/Shared/Intrinsics.cs
… src/VCDiff/Shared/NativeAllocation.cs
… src/VCDiff/Shared/PinnedArrayRental.cs
… src/VCDiff/Shared/Pool.cs
… src/VCDiff/Shared/VarIntBE.cs
… src/VCDiff/VCDiff.csproj
… src/VCDiff/VCDiff.xml