@chyyran 2021-11-23
LGTM. Will push a release soon so you can use mainline on NuGet. Apologies again for the late reply.