@chyyran 2021-11-23
Not indexed yet but published as https://www.nuget.org/packages/VCDiff/4.0.0

Capping `bufferSize` is technically a breaking change + wouldn't want to accidentally break other consumers, so I bumped major version just in case.